requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
moviepy==1.0.3
pillow==11.0.0
python-dotenv==1.0.1
pandas==2.2.3
requests==2.32.3
imageio==2.36.1
